Moved some classes to Control directory with namespace Xamarin.Forms.Controls to fix conflicts after namespaces changes
