+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.Shared/Bugzilla32691.cs
@@ -9,11 +9,6 @@ using NUnit.Framework;
 
 namespace Xamarin.Forms.Controls.Issues
 {
-	public interface ICacheService
-	{
-		void ClearImageCache ();
-	}
-
 	[Preserve (AllMembers = true)]
 	[Issue (IssueTracker.Bugzilla, 32691, "Clearing an image by setting Image.Source to null, while Image.IsLoading is true, doesn't work.")]
 	public class Bugzilla32691 : TestContentPage
