+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.Shared/Controls/ICacheService.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Xamarin.Forms.Controls
+{
+	public interface ICacheService
+	{
+		void ClearImageCache ();
+	}
+}
